+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Changed the beer-log home tile to show minutes/hours instead of "0 days ago" for logs within the same day (#69)
+
 ## [1.0.0] - 2026-09-15
 
 ### Added